refactor(api-v1): split group state route registration
