+++ apps/api-v1/test/group-state/group-admission-routes.test.ts
@@ -1,22 +1,12 @@
+// deno-fmt-ignore-file
 import assert from 'node:assert/strict';
-
 import { Either } from '@shared/resilience/Either.ts';
-import type { AppInboxEnqueueInput } from '@shared-server/rallar-system/services/AppInboxService.ts';
+import { type AppInboxEnqueueInput, AppInboxType } from '@shared-server/rallar-system/services/AppInboxService.ts';
 import { toGroupStateCommand } from '../../src/group-state/to-group-state-command.ts';
 import { toGroupStateResponse } from '../../src/group-state/to-group-state-response.ts';
-import * as groupStateRoutes from '../../src/routes/group-state-routes.ts';
-
-import {
-  createGroupStateRouteAuthSession,
-  createGroupStateRouteEvent,
-  createGroupStateRouteSnapshot,
-  createGroupStateRouteTestRuntime,
-  TEST_GROUP_SCOPE,
-  toGroupStateWritten,
-} from './group-state-route-test-runtime.ts';
-
+import type { GroupStateRouteAuthSession, ProcessGroupAppInbox } from '../../src/group-state/group-state-route-contracts.ts';
+import { createGroupStateRouteAuthSession, createGroupStateRouteEvent, createGroupStateRouteSnapshot, createGroupStateRouteTestDependencies, createGroupStateRouteTestRuntime, TEST_GROUP_SCOPE, toGroupStateWritten, withStrictGroupStateRouteReadAuth } from './group-state-route-test-runtime.ts';
 const API_BASE = '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1';
-
 Deno.test('group admission commands retain every authenticated AppInbox envelope', () => {
   const authSession = createGroupStateRouteAuthSession('alice');
   const commands = [
@@ -83,13 +73,11 @@ Deno.test('group admission commands retain every authenticated AppInbox envelope
       },
     }),
   ];
-
   assert.equal(
     JSON.stringify(commands),
     '[{"type":"GROUP_JOIN","resourceId":"join-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","request":{"inviteToken":"invite-1","joinCode":"code-1","actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"join-request"}}},{"type":"GROUP_INVITE_ACCEPT","resourceId":"accept-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","request":{"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"accept-request"}}},{"type":"GROUP_JOIN_CODE_ROTATE","resourceId":"rotate-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","request":{"joinCode":"next-code","expiresAtEpochMs":2000,"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"rotate-request"}}},{"type":"GROUP_INVITE_CREATE","resourceId":"invite-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","principalId":"bob","request":{"invitationExpiresAtEpochMs":2000,"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"invite-request"}}},{"type":"GROUP_INVITE_REVOKE","resourceId":"revoke-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","principalId":"bob","request":{"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"revoke-request"}}}]',
   );
 });
-
 Deno.test('group join-code response omits its event without changing JSON order', () => {
   const snapshot = createGroupStateRouteSnapshot('room-1');
   const response = toGroupStateResponse({
@@ -104,21 +92,18 @@ Deno.test('group join-code response omits its event without changing JSON order'
       }),
     },
   });
-
   assert.strictEqual(response.snapshot, snapshot);
   assert.equal(
     JSON.stringify(response),
     '{"joinCode":"next-code","expiresAtEpochMs":2000,"snapshot":{"stateRevision":1,"causalRevision":{"groupRevision":1,"presenceRevision":0},"group":{"applicationId":"app-1","workspaceId":"workspace-1","groupId":"room-1","slug":null,"displayName":"room-1","description":null,"kind":"room","status":"active","joinMode":"open","maxMembers":null,"maxSessionsPerMember":null,"metadata":{},"activeMemberCount":1,"ownerPrincipalId":"alice","snapshotVersion":1,"metadataVersion":1,"rosterVersion":1,"presenceVersion":0,"created":{"atEpochMs":1,"actor":{"kind":"service","serviceId":"test"},"reason":null,"traceId":null,"requestId":null},"updated":{"atEpochMs":1,"actor":{"kind":"service","serviceId":"test"},"reason":null,"traceId":null,"requestId":null},"expiresAtEpochMs":null,"emptySinceEpochMs":null,"purgeAfterEpochMs":null,"archived":null,"deleted":null},"members":[{"applicationId":"app-1","workspaceId":"workspace-1","groupId":"room-1","principalId":"alice","role":"owner","status":"active","joined":{"atEpochMs":1,"actor":{"kind":"service","serviceId":"test"},"reason":null,"traceId":null,"requestId":null},"updated":{"atEpochMs":1,"actor":{"kind":"service","serviceId":"test"},"reason":null,"traceId":null,"requestId":null},"left":null,"removed":null,"banned":null,"invitedByPrincipalId":null,"invitationExpiresAtEpochMs":null}],"activeSessions":[],"memberCount":1,"onlineMemberCount":0}}',
   );
 });
-
 Deno.test('group admission routes retain every AppInbox envelope and actor override', async () => {
   const enqueued: unknown[] = [];
   const snapshot = createGroupStateRouteSnapshot('room-1');
   const runtime = createGroupStateRouteTestRuntime({
     processGroupAppInbox: captureGroupStateWrite(enqueued, snapshot),
   });
-
   const responses = [
     await requestGroupMutation(runtime.app, `${API_BASE}/join`, {
       inviteToken: 'invite-1',
@@ -151,7 +136,6 @@ Deno.test('group admission routes retain every AppInbox envelope and actor overr
       requestId: 'revoke-request',
     }),
   ];
-
   for (const response of responses) {
     assert.equal(response.status, 200);
   }
@@ -160,20 +144,12 @@ Deno.test('group admission routes retain every AppInbox envelope and actor overr
     '[{"type":"GROUP_JOIN","resourceId":"join-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","request":{"inviteToken":"invite-1","joinCode":"code-1","actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"join-request"}}},{"type":"GROUP_INVITE_ACCEPT","resourceId":"accept-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","request":{"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"accept-request"}}},{"type":"GROUP_JOIN_CODE_ROTATE","resourceId":"rotate-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","request":{"joinCode":"next-code","expiresAtEpochMs":2000,"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"rotate-request"}}},{"type":"GROUP_INVITE_CREATE","resourceId":"invite-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","principalId":"bob","request":{"invitationExpiresAtEpochMs":2000,"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"invite-request"}}},{"type":"GROUP_INVITE_REVOKE","resourceId":"revoke-request","contextId":"app-1:workspace-1:room-1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-1","principalId":"bob","request":{"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"revoke-request"}}}]',
   );
 });
-
-function captureGroupStateWrite(
-  enqueued: unknown[],
-  snapshot: ReturnType<typeof createGroupStateRouteSnapshot>,
-): groupStateRoutes.ProcessGroupAppInbox {
-  return <V, R>(
-    _authority: groupStateRoutes.GroupStateRouteAuthSession,
-    entry: AppInboxEnqueueInput<V>,
-  ): Promise<R> => {
+function captureGroupStateWrite(enqueued: unknown[], snapshot: ReturnType<typeof createGroupStateRouteSnapshot>): ProcessGroupAppInbox {
+  return <V, R>(_authority: GroupStateRouteAuthSession, entry: AppInboxEnqueueInput<V>): Promise<R> => {
     enqueued.push(entry);
     return Promise.resolve(toGroupStateWritten(snapshot) as R);
   };
 }
-
 async function requestGroupMutation(
   app: ReturnType<typeof createGroupStateRouteTestRuntime>['app'],
   path: string,
@@ -185,3 +161,238 @@ async function requestGroupMutation(
     body: JSON.stringify(body),
   });
 }
+Deno.test('group join route enqueues explicit join intent with authenticated actor', async () => {
+  await withStrictGroupStateRouteReadAuth(false, async () => {
+    const snapshot = createGroupStateRouteSnapshot('room-1', ['alice']);
+    const enqueued: AppInboxEnqueueInput<unknown>[] = [];
+    const { app } = createGroupStateRouteTestRuntime({
+      session: createGroupStateRouteAuthSession('alice'),
+      groupService: {},
+      processGroupAppInbox: <V, R>(
+        _authority: GroupStateRouteAuthSession,
+        input: AppInboxEnqueueInput<V>,
+      ): Promise<R> => {
+        enqueued.push(input);
+        return Promise.resolve({
+          status: 'ok',
+          result: {
+            right: {
+              snapshot,
+            },
+          },
+        } as R);
+      },
+    });
+    const response = await app.request(
+      '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/join',
+      {
+        method: 'POST',
+        headers: {
+          authorization: 'Bearer token',
+          'content-type': 'application/json',
+        },
+        body: JSON.stringify({
+          inviteToken: 'invite-1',
+          joinCode: 'code-1',
+          requestId: 'join-request-1',
+        }),
+      },
+    );
+    assert.equal(response.status, 200);
+    assert.deepEqual(await response.json(), snapshot);
+    assert.equal(enqueued.length, 1);
+    assert.deepEqual(enqueued[0], {
+      type: AppInboxType.GROUP_JOIN,
+      resourceId: 'join-request-1',
+      contextId: 'app-1:workspace-1:room-1',
+      senderId: 'alice',
+      data: {
+        scope: TEST_GROUP_SCOPE,
+        groupId: 'room-1',
+        request: {
+          inviteToken: 'invite-1',
+          joinCode: 'code-1',
+          actorPrincipalId: 'alice',
+          actorSessionId: 'alice-session',
+          requestId: 'join-request-1',
+        },
+      },
+    });
+  });
+});
+Deno.test('group invite routes enqueue safe invite workflows with authenticated actors', async () => {
+  await withStrictGroupStateRouteReadAuth(false, async () => {
+    const snapshot = createGroupStateRouteSnapshot('room-1', ['alice']);
+    const enqueued: AppInboxEnqueueInput<unknown>[] = [];
+    const { app } = createGroupStateRouteTestRuntime({
+      session: createGroupStateRouteAuthSession('alice'),
+      groupService: {},
+      processGroupAppInbox: <V, R>(
+        _authority: GroupStateRouteAuthSession,
+        input: AppInboxEnqueueInput<V>,
+      ): Promise<R> => {
+        enqueued.push(input);
+        return Promise.resolve({
+          status: 'ok',
+          result: {
+            right: {
+              snapshot,
+            },
+          },
+        } as R);
+      },
+    });
+    const createResponse = await app.request(
+      '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/invites/bob',
+      {
+        method: 'POST',
+        headers: {
+          authorization: 'Bearer token',
+          'content-type': 'application/json',
+        },
+        body: JSON.stringify({
+          invitationExpiresAtEpochMs: 2_000,
+          requestId: 'invite-create-1',
+        }),
+      },
+    );
+    const revokeResponse = await app.request(
+      '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/invites/bob/revoke',
+      {
+        method: 'POST',
+        headers: {
+          authorization: 'Bearer token',
+          'content-type': 'application/json',
+        },
+        body: JSON.stringify({ requestId: 'invite-revoke-1' }),
+      },
+    );
+    const acceptResponse = await app.request(
+      '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/invites/accept',
+      {
+        method: 'POST',
+        headers: {
+          authorization: 'Bearer token',
+          'content-type': 'application/json',
+        },
+        body: JSON.stringify({ requestId: 'invite-accept-1' }),
+      },
+    );
+    assert.equal(createResponse.status, 200);
+    assert.equal(revokeResponse.status, 200);
+    assert.equal(acceptResponse.status, 200);
+    assert.deepEqual(enqueued, [
+      {
+        type: AppInboxType.GROUP_INVITE_CREATE,
+        resourceId: 'invite-create-1',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          principalId: 'bob',
+          request: {
+            invitationExpiresAtEpochMs: 2_000,
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'invite-create-1',
+          },
+        },
+      },
+      {
+        type: AppInboxType.GROUP_INVITE_REVOKE,
+        resourceId: 'invite-revoke-1',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          principalId: 'bob',
+          request: {
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'invite-revoke-1',
+          },
+        },
+      },
+      {
+        type: AppInboxType.GROUP_INVITE_ACCEPT,
+        resourceId: 'invite-accept-1',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          request: {
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'invite-accept-1',
+          },
+        },
+      },
+    ]);
+  });
+});
+Deno.test('group join-code route enqueues rotation workflow with authenticated actor', async () => {
+  await withStrictGroupStateRouteReadAuth(false, async () => {
+    const snapshot = createGroupStateRouteSnapshot('room-1', ['alice']);
+    const responseBody = {
+      joinCode: 'code-1',
+      expiresAtEpochMs: 2_000,
+      snapshot,
+    };
+    const enqueued: AppInboxEnqueueInput<unknown>[] = [];
+    const { app } = createGroupStateRouteTestRuntime({
+      session: createGroupStateRouteAuthSession('alice'),
+      groupService: {},
+      processGroupAppInbox: <V, R>(
+        _authority: GroupStateRouteAuthSession,
+        input: AppInboxEnqueueInput<V>,
+      ): Promise<R> => {
+        enqueued.push(input);
+        return Promise.resolve({
+          status: 'ok',
+          result: {
+            right: responseBody,
+          },
+        } as R);
+      },
+    });
+    const response = await app.request(
+      '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/join-code/rotate',
+      {
+        method: 'POST',
+        headers: {
+          authorization: 'Bearer token',
+          'content-type': 'application/json',
+        },
+        body: JSON.stringify({
+          joinCode: 'code-1',
+          expiresAtEpochMs: 2_000,
+          requestId: 'rotate-code-1',
+        }),
+      },
+    );
+    assert.equal(response.status, 200);
+    assert.deepEqual(await response.json(), responseBody);
+    assert.deepEqual(enqueued, [
+      {
+        type: AppInboxType.GROUP_JOIN_CODE_ROTATE,
+        resourceId: 'rotate-code-1',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          request: {
+            joinCode: 'code-1',
+            expiresAtEpochMs: 2_000,
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'rotate-code-1',
+          },
+        },
+      },
+    ]);
+  });
+});
