+++ apps/api-v1/test/group-state/group-membership-routes.test.ts
@@ -1,15 +1,23 @@
 import assert from 'node:assert/strict';
 
-import type { AppInboxEnqueueInput } from '@shared-server/rallar-system/services/AppInboxService.ts';
+import {
+  type AppInboxEnqueueInput,
+  AppInboxType,
+} from '@shared-server/rallar-system/services/AppInboxService.ts';
 import { toGroupStateCommand } from '../../src/group-state/to-group-state-command.ts';
-import * as groupStateRoutes from '../../src/routes/group-state-routes.ts';
+import type {
+  GroupStateRouteAuthSession,
+  ProcessGroupAppInbox,
+} from '../../src/group-state/group-state-route-contracts.ts';
 
 import {
   createGroupStateRouteAuthSession,
   createGroupStateRouteSnapshot,
+  createGroupStateRouteTestDependencies,
   createGroupStateRouteTestRuntime,
   TEST_GROUP_SCOPE,
   toGroupStateWritten,
+  withStrictGroupStateRouteReadAuth,
 } from './group-state-route-test-runtime.ts';
 
 const API_BASE = '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1';
@@ -156,9 +164,9 @@ Deno.test('group membership routes retain every AppInbox envelope and self-servi
 function captureGroupStateWrite(
   enqueued: unknown[],
   snapshot: ReturnType<typeof createGroupStateRouteSnapshot>,
-): groupStateRoutes.ProcessGroupAppInbox {
+): ProcessGroupAppInbox {
   return <V, R>(
-    _authority: groupStateRoutes.GroupStateRouteAuthSession,
+    _authority: GroupStateRouteAuthSession,
     entry: AppInboxEnqueueInput<V>,
   ): Promise<R> => {
     enqueued.push(entry);
@@ -178,3 +186,176 @@ async function requestGroupMutation(
     body: JSON.stringify(body),
   });
 }
+
+Deno.test('group governance routes enqueue safe workflows with authenticated actors', async () => {
+  await withStrictGroupStateRouteReadAuth(false, async () => {
+    const snapshot = createGroupStateRouteSnapshot('room-1', ['alice', 'bob']);
+    const enqueued: AppInboxEnqueueInput<unknown>[] = [];
+    const { app } = createGroupStateRouteTestRuntime({
+      session: createGroupStateRouteAuthSession('alice'),
+      groupService: {},
+      processGroupAppInbox: <V, R>(
+        _authority: GroupStateRouteAuthSession,
+        input: AppInboxEnqueueInput<V>,
+      ): Promise<R> => {
+        enqueued.push(input);
+        return Promise.resolve({
+          status: 'ok',
+          result: {
+            right: { snapshot },
+          },
+        } as R);
+      },
+    });
+
+    const requests = [
+      app.request(
+        '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/members/bob/remove',
+        {
+          method: 'POST',
+          headers: {
+            authorization: 'Bearer token',
+            'content-type': 'application/json',
+          },
+          body: JSON.stringify({ requestId: 'remove-bob' }),
+        },
+      ),
+      app.request(
+        '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/members/bob/ban',
+        {
+          method: 'POST',
+          headers: {
+            authorization: 'Bearer token',
+            'content-type': 'application/json',
+          },
+          body: JSON.stringify({ requestId: 'ban-bob' }),
+        },
+      ),
+      app.request(
+        '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/members/bob/unban',
+        {
+          method: 'POST',
+          headers: {
+            authorization: 'Bearer token',
+            'content-type': 'application/json',
+          },
+          body: JSON.stringify({ requestId: 'unban-bob' }),
+        },
+      ),
+      app.request(
+        '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/members/bob/role',
+        {
+          method: 'PUT',
+          headers: {
+            authorization: 'Bearer token',
+            'content-type': 'application/json',
+          },
+          body: JSON.stringify({ role: 'admin', requestId: 'role-bob' }),
+        },
+      ),
+      app.request(
+        '/api/state/apps/app-1/workspaces/workspace-1/groups/room-1/owner/transfer',
+        {
+          method: 'POST',
+          headers: {
+            authorization: 'Bearer token',
+            'content-type': 'application/json',
+          },
+          body: JSON.stringify({
+            newOwnerPrincipalId: 'bob',
+            requestId: 'transfer-owner',
+          }),
+        },
+      ),
+    ];
+    const responses = await Promise.all(requests);
+
+    for (const response of responses) {
+      assert.equal(response.status, 200);
+      assert.deepEqual(await response.json(), snapshot);
+    }
+    assert.deepEqual(enqueued, [
+      {
+        type: AppInboxType.GROUP_MEMBER_REMOVE,
+        resourceId: 'remove-bob',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          principalId: 'bob',
+          request: {
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'remove-bob',
+          },
+        },
+      },
+      {
+        type: AppInboxType.GROUP_MEMBER_BAN,
+        resourceId: 'ban-bob',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          principalId: 'bob',
+          request: {
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'ban-bob',
+          },
+        },
+      },
+      {
+        type: AppInboxType.GROUP_MEMBER_UNBAN,
+        resourceId: 'unban-bob',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          principalId: 'bob',
+          request: {
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'unban-bob',
+          },
+        },
+      },
+      {
+        type: AppInboxType.GROUP_MEMBER_ROLE_SET,
+        resourceId: 'role-bob',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          principalId: 'bob',
+          request: {
+            role: 'admin',
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'role-bob',
+          },
+        },
+      },
+      {
+        type: AppInboxType.GROUP_OWNERSHIP_TRANSFER,
+        resourceId: 'transfer-owner',
+        contextId: 'app-1:workspace-1:room-1',
+        senderId: 'alice',
+        data: {
+          scope: TEST_GROUP_SCOPE,
+          groupId: 'room-1',
+          request: {
+            newOwnerPrincipalId: 'bob',
+            actorPrincipalId: 'alice',
+            actorSessionId: 'alice-session',
+            requestId: 'transfer-owner',
+          },
+        },
+      },
+    ]);
+  });
+});
