+++ apps/api-v1/test/group-state/group-state-mutation-routes.test.ts
@@ -1,37 +1,20 @@
+// deno-fmt-ignore-file
 import assert from 'node:assert/strict';
-
+import type { GroupSnapshot } from '@shared/api/group-types.ts';
 import { Either } from '@shared/resilience/Either.ts';
-import type {
-  GroupCreateAppInboxPayload,
-  GroupUpdateAppInboxPayload,
-} from '@shared-server/rallar-system/services/AppGroupInboxService.ts';
-import {
-  type AppInboxEnqueueInput,
-  AppInboxType,
-} from '@shared-server/rallar-system/services/AppInboxService.ts';
-
-import {
-  readGroupStateRouteRequest,
-} from '../../src/group-state/read-group-state-route-request.ts';
+import type { GroupCreateAppInboxPayload, GroupUpdateAppInboxPayload } from '@shared-server/rallar-system/services/AppGroupInboxService.ts';
+import { type AppInboxEnqueueInput, AppInboxType } from '@shared-server/rallar-system/services/AppInboxService.ts';
+import { readGroupStateRouteRequest } from '../../src/group-state/read-group-state-route-request.ts';
 import { toGroupStateCommand } from '../../src/group-state/to-group-state-command.ts';
 import { toGroupStateResponse } from '../../src/group-state/to-group-state-response.ts';
-import * as groupStateRoutes from '../../src/routes/group-state-routes.ts';
-
-import {
-  createGroupStateRouteAuthSession,
-  createGroupStateRouteSnapshot,
-  createGroupStateRouteTestRuntime,
-  TEST_GROUP_SCOPE,
-  toGroupStateWritten,
-} from './group-state-route-test-runtime.ts';
-
+import type { GroupStateRouteAuthSession, ProcessGroupAppInbox } from '../../src/group-state/group-state-route-contracts.ts';
+import { createGroupStateRouteAuthSession, createGroupStateRouteSnapshot, createGroupStateRouteTestDependencies, createGroupStateRouteTestRuntime, TEST_GROUP_SCOPE, toGroupStateWritten, withStrictGroupStateRouteReadAuth } from './group-state-route-test-runtime.ts';
 const API_BASE = '/api/state/apps/app-1/workspaces/workspace-1/groups';
 const RANDOM_UUID_DESCRIPTOR_AT_MODULE_LOAD = Object.getOwnPropertyDescriptor(
   crypto,
   'randomUUID',
 );
 const RANDOM_UUID_AT_MODULE_LOAD = crypto.randomUUID;
-
 Deno.test('canonical group request reader retains body request ID precedence', async () => {
   const request = await readGroupStateRouteRequest<{ requestId?: string; name: string }>({
     req: {
@@ -39,10 +22,8 @@ Deno.test('canonical group request reader retains body request ID precedence', a
       header: () => 'header-request',
     },
   });
-
   assert.deepEqual(request, { requestId: 'body-request', name: 'Room' });
 });
-
 Deno.test('group create command retains its authenticated AppInbox envelope', () => {
   const command = toGroupStateCommand({
     operation: 'create-group',
@@ -58,13 +39,11 @@ Deno.test('group create command retains its authenticated AppInbox envelope', ()
       requestId: 'create-body',
     },
   });
-
   assert.equal(
     JSON.stringify(command),
     '{"type":"GROUP_CREATE","resourceId":"create-body","contextId":"app-1:workspace-1:room%2F1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"request":{"groupId":"room/1","displayName":"Room","kind":"room","createdByPrincipalId":"alice","actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"create-body"}}}',
   );
 });
-
 Deno.test('group command output keeps create payloads isolated from updates', () => {
   const command = toGroupStateCommand({
     operation: 'create-group',
@@ -78,7 +57,6 @@ Deno.test('group command output keeps create payloads isolated from updates', ()
       requestId: 'create-request',
     },
   });
-
   assert.equal(command.type, AppInboxType.GROUP_CREATE);
   if (command.type === AppInboxType.GROUP_CREATE) {
     const payload: GroupCreateAppInboxPayload = command.data;
@@ -88,7 +66,6 @@ Deno.test('group command output keeps create payloads isolated from updates', ()
     void invalidPayload;
   }
 });
-
 Deno.test('group aggregate commands retain update and director envelopes', () => {
   const authSession = createGroupStateRouteAuthSession('alice');
   const commands = [
@@ -117,20 +94,17 @@ Deno.test('group aggregate commands retain update and director envelopes', () =>
       },
     }),
   ];
-
   assert.equal(
     JSON.stringify(commands),
     '[{"type":"GROUP_UPDATE","resourceId":"update-body","contextId":"app-1:workspace-1:room-2","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-2","request":{"displayName":"Renamed","actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"update-body"}}},{"type":"GROUP_DIRECTOR_APPOINT","resourceId":"appoint-body","contextId":"app-1:workspace-1:room-3","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-3","request":{"heartbeatTtlMs":20,"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"appoint-body"}}}]',
   );
 });
-
 Deno.test('group mutation response retains snapshot identity and durable error text', () => {
   const snapshot = createGroupStateRouteSnapshot('room-1');
   const response = toGroupStateResponse({
     kind: 'mutation',
     written: toGroupStateWritten(snapshot),
   });
-
   assert.strictEqual(response.snapshot, snapshot);
   assert.equal(
     JSON.stringify(response.snapshot),
@@ -145,7 +119,6 @@ Deno.test('group mutation response retains snapshot identity and durable error t
     { message: 'Mutation result rejected' },
   );
 });
-
 Deno.test('group aggregate routes retain their AppInbox envelopes', async () => {
   const enqueued: unknown[] = [];
   const snapshot = createGroupStateRouteSnapshot('room-1');
@@ -153,7 +126,6 @@ Deno.test('group aggregate routes retain their AppInbox envelopes', async () => 
     groupService: { readSnapshot: () => Promise.resolve(snapshot) },
     processGroupAppInbox: captureGroupStateWrite(enqueued, snapshot),
   });
-
   const responses = [
     await requestGroupMutation(runtime.app, API_BASE, 'POST', {
       groupId: 'room/1',
@@ -177,7 +149,6 @@ Deno.test('group aggregate routes retain their AppInbox envelopes', async () => 
       requestId: 'appoint-body',
     }),
   ];
-
   assert.equal(responses[0].status, 201);
   assert.equal(responses[1].status, 200);
   assert.equal(responses[2].status, 200);
@@ -186,7 +157,6 @@ Deno.test('group aggregate routes retain their AppInbox envelopes', async () => 
     '[{"type":"GROUP_CREATE","resourceId":"create-body","contextId":"app-1:workspace-1:room%2F1","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"request":{"groupId":"room/1","displayName":"Room","kind":"room","createdByPrincipalId":"alice","actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"create-body"}}},{"type":"GROUP_UPDATE","resourceId":"update-body","contextId":"app-1:workspace-1:room-2","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-2","request":{"displayName":"Renamed","actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"update-body"}}},{"type":"GROUP_DIRECTOR_APPOINT","resourceId":"appoint-body","contextId":"app-1:workspace-1:room-3","senderId":"alice","data":{"scope":{"applicationId":"app-1","workspaceId":"workspace-1"},"groupId":"room-3","request":{"heartbeatTtlMs":20,"actorPrincipalId":"alice","actorSessionId":"alice-session","requestId":"appoint-body"}}}]',
   );
 });
-
 Deno.test(
   'group aggregate routes preserve body, header, then one generated request ID',
   async () => {
@@ -195,7 +165,6 @@ Deno.test(
     const runtime = createGroupStateRouteTestRuntime({
       processGroupAppInbox: captureGroupStateWrite(enqueued, snapshot),
     });
-
     await withRandomUuid('generated-request', async (readRandomCallCount) => {
       const bodyResponse = await requestGroupMutation(runtime.app, API_BASE, 'POST', {
         groupId: 'body-id-group',
@@ -213,13 +182,11 @@ Deno.test(
         displayName: 'Generated',
         kind: 'room',
       });
-
       assert.equal(bodyResponse.status, 201);
       assert.equal(headerResponse.status, 201);
       assert.equal(generatedResponse.status, 201);
       assert.equal(readRandomCallCount(), 1);
     });
-
     assert.equal(
       JSON.stringify(enqueued.map((entry) => {
         const envelope = entry as AppInboxEnqueueInput<unknown>;
@@ -229,7 +196,6 @@ Deno.test(
     );
   },
 );
-
 Deno.test(
   'group aggregate request ID UUID stub restores crypto randomUUID observable shape',
   () => {
@@ -241,7 +207,6 @@ Deno.test(
     assert.notEqual(crypto.randomUUID(), 'generated-request');
   },
 );
-
 Deno.test(
   'group aggregate route waits for AppInbox completion before its normal response',
   async () => {
@@ -250,7 +215,7 @@ Deno.test(
     const snapshot = createGroupStateRouteSnapshot('room-1');
     const runtime = createGroupStateRouteTestRuntime({
       processGroupAppInbox: <V, R>(
-        _authority: groupStateRoutes.GroupStateRouteAuthSession,
+        _authority: GroupStateRouteAuthSession,
         _entry: AppInboxEnqueueInput<V>,
       ): Promise<R> =>
         new Promise((resolve) => {
@@ -258,7 +223,6 @@ Deno.test(
           resolveCompletion = () => resolve(toGroupStateWritten(snapshot) as R);
         }),
     });
-
     const responsePromise = requestGroupMutation(runtime.app, API_BASE, 'POST', {
       groupId: 'room-1',
       displayName: 'Room',
@@ -266,17 +230,14 @@ Deno.test(
       requestId: 'await-completion',
     });
     await new Promise((resolve) => setTimeout(resolve, 0));
-
     assert.equal(enqueued, 1);
     assert.ok(resolveCompletion);
     resolveCompletion();
-
     const response = await responsePromise;
     assert.equal(response.status, 201);
     assert.deepEqual(await response.json(), snapshot);
   },
 );
-
 Deno.test(
   'group aggregate route stops at route authentication failure before AppInbox',
   async () => {
@@ -289,19 +250,16 @@ Deno.test(
         return Promise.resolve(undefined as never);
       },
     });
-
     const response = await requestGroupMutation(runtime.app, API_BASE, 'POST', {
       groupId: 'room-1',
       displayName: 'Room',
       kind: 'room',
     });
-
     assert.equal(response.status, 400);
     assert.deepEqual(await response.json(), { error: 'route authentication failed' });
     assert.equal(enqueued, 0);
   },
 );
-
 Deno.test(
   'group aggregate route serializes AppInbox failure after the awaited completion',
   async () => {
@@ -313,13 +271,11 @@ Deno.test(
     const runtime = createGroupStateRouteTestRuntime({
       processGroupAppInbox: () => Promise.reject(failure),
     });
-
     const response = await requestGroupMutation(runtime.app, API_BASE, 'POST', {
       groupId: 'room-1',
       displayName: 'Room',
       kind: 'room',
     });
-
     assert.equal(response.status, 409);
     assert.equal(
       JSON.stringify(await response.json()),
@@ -327,20 +283,18 @@ Deno.test(
     );
   },
 );
-
 function captureGroupStateWrite(
   enqueued: unknown[],
   snapshot: ReturnType<typeof createGroupStateRouteSnapshot>,
-): groupStateRoutes.ProcessGroupAppInbox {
+): ProcessGroupAppInbox {
   return <V, R>(
-    _authority: groupStateRoutes.GroupStateRouteAuthSession,
+    _authority: GroupStateRouteAuthSession,
     entry: AppInboxEnqueueInput<V>,
   ): Promise<R> => {
     enqueued.push(entry);
     return Promise.resolve(toGroupStateWritten(snapshot) as R);
   };
 }
-
 async function requestGroupMutation(
   app: ReturnType<typeof createGroupStateRouteTestRuntime>['app'],
   path: string,
@@ -354,7 +308,6 @@ async function requestGroupMutation(
     body: JSON.stringify(body),
   });
 }
-
 async function withRandomUuid(
   value: string,
   action: (readCallCount: () => number) => Promise<void>,
@@ -378,3 +331,64 @@ async function withRandomUuid(
     }
   }
 }
+Deno.test('all non-presence group REST mutations reject malformed bodies before inbox enqueue', async () => {
+  const processCalls: unknown[] = [];
+  const snapshot = createGroupStateRouteSnapshot('room-1', ['alice']);
+  const ownerSnapshot: GroupSnapshot = {
+    ...snapshot,
+    members: snapshot.members.map((member) => ({ ...member, role: 'owner' as const })),
+  };
+  const { app } = createGroupStateRouteTestRuntime({
+    session: createGroupStateRouteAuthSession('alice'),
+    groupService: {
+      readSnapshot: () => Promise.resolve(ownerSnapshot),
+    },
+    processGroupAppInbox: (_authority, input) => {
+      processCalls.push(input);
+      return Promise.reject(new Error('Malformed request reached group inbox'));
+    },
+  });
+  const base = '/api/state/apps/app-1/workspaces/workspace-1/groups';
+  const group = `${base}/room-1`;
+  const cases = [
+    { method: 'POST', path: base, body: { displayName: 7, kind: 'room', groupId: 'room-2' } },
+    { method: 'PUT', path: group, body: { status: 'unknown' } },
+    { method: 'POST', path: `${group}/director/appoint`, body: { heartbeatTtlMs: 0 } },
+    { method: 'POST', path: `${group}/join`, body: { inviteToken: 7 } },
+    { method: 'POST', path: `${group}/invites/accept`, body: { reason: 7 } },
+    {
+      method: 'POST',
+      path: `${group}/join-code/rotate`,
+      body: { joinCode: '', expiresAtEpochMs: 0 },
+    },
+    {
+      method: 'POST',
+      path: `${group}/invites/bob`,
+      body: { invitationExpiresAtEpochMs: -1 },
+    },
+    { method: 'POST', path: `${group}/invites/bob/revoke`, body: { traceId: 7 } },
+    { method: 'POST', path: `${group}/members/bob/remove`, body: { reason: {} } },
+    { method: 'POST', path: `${group}/members/bob/ban`, body: { requestId: {} } },
+    { method: 'POST', path: `${group}/members/bob/unban`, body: { traceId: [] } },
+    { method: 'PUT', path: `${group}/members/bob/role`, body: { role: 'superuser' } },
+    { method: 'POST', path: `${group}/owner/transfer`, body: { newOwnerPrincipalId: '' } },
+    {
+      method: 'PUT',
+      path: `${group}/members/alice`,
+      body: { status: 'active', invitationExpiresAtEpochMs: -1 },
+    },
+  ] as const;
+  for (const testCase of cases) {
+    const response = await app.request(testCase.path, {
+      method: testCase.method,
+      headers: {
+        authorization: 'Bearer token',
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify(testCase.body),
+    });
+    assert.equal(response.status, 400, `${testCase.method} ${testCase.path}`);
+    assert.match((await response.json()).error, /Group|group/);
+  }
+  assert.equal(processCalls.length, 0);
+});
